+++ TLM/TLM/Custom/PathFinding/CustomPathFind.cs
@@ -65,6 +65,8 @@ namespace TrafficManager.Custom.PathFinding {
 
         private GlobalConfig globalConf_;
 
+        private SavedGameOptions savedGameOptions_;
+
         private struct BufferItem {
             public PathUnit.Position Position;
             public float ComparisonValue;
@@ -140,6 +142,11 @@ namespace TrafficManager.Custom.PathFinding {
 
         // custom fields
         private PathUnitQueueItem queueItem_;
+        /// <summary>
+        /// Contains allowed extVehicle types, in normal conditions just QueueIten.vehicleType,
+        /// when path may include CargoVehicle lanes then it contains also respective cargo vehicle types
+        /// </summary>
+        private ExtVehicleType allowedPathVehicleTypes_;
         private bool isHeavyVehicle_;
 
         private bool debugLog_; // this will be false in non-Debug
@@ -252,7 +259,9 @@ namespace TrafficManager.Custom.PathFinding {
         }
 
         private void PathFindImplementation(uint unit, ref PathUnit data) {
+
             globalConf_ = GlobalConfig.Instance; // NON-STOCK CODE
+            savedGameOptions_ = SavedGameOptions.Instance;
 
             NetManager netManager = Singleton<NetManager>.instance;
 
@@ -289,8 +298,15 @@ namespace TrafficManager.Custom.PathFinding {
                 disableMask_ |= NetSegment.Flags.Flooded;
             }
 
+            allowedPathVehicleTypes_ = queueItem_.vehicleType;
             if ((laneTypes_ & NetInfo.LaneType.Vehicle) != NetInfo.LaneType.None) {
                 laneTypes_ |= NetInfo.LaneType.TransportVehicle;
+
+                // allow mixed vehicle path when requested path is Vehicle + CargoVehicle (e.g.: CargoTruck and PostVanAI)
+                bool allowMixedCargoPath = (laneTypes_ & NetInfo.LaneType.CargoVehicle) != NetInfo.LaneType.None;
+                if (allowMixedCargoPath) {
+                    allowedPathVehicleTypes_ |= VehicleTypesToMixedCargoExtVehicle(vehicleTypes_);
+                }
             }
 
 #if ROUTING
@@ -298,11 +314,11 @@ namespace TrafficManager.Custom.PathFinding {
                 (queueItem_.vehicleType & ExtVehicleType.RoadVehicle) != ExtVehicleType.None;
 
             isLaneArrowObeyingEntity_ =
-                (!Options.relaxedBusses || queueItem_.vehicleType != ExtVehicleType.Bus) &&
+                (!savedGameOptions_.relaxedBusses || queueItem_.vehicleType != ExtVehicleType.Bus) &&
                 (vehicleTypes_ & LaneArrowManager.VEHICLE_TYPES) != VehicleInfo.VehicleType.None &&
                 (queueItem_.vehicleType & LaneArrowManager.EXT_VEHICLE_TYPES) != ExtVehicleType.None;
 #if DEBUG
-            isLaneArrowObeyingEntity_ &= !Options.allRelaxed;
+            isLaneArrowObeyingEntity_ &= !savedGameOptions_.allRelaxed;
 #endif
 #endif
 
@@ -1263,7 +1279,7 @@ namespace TrafficManager.Custom.PathFinding {
 
                         if (isPedZoneRoad &&
                             // NON-STOCK CODE START skip execution when driving-only path for passenger car
-                            !(Options.parkingAI &&
+                            !(savedGameOptions_.parkingAI &&
                               queueItem_.vehicleType == ExtVehicleType.PassengerCar &&
                               queueItem_.pathType == ExtPathType.DrivingOnly)
                             // NON-STOCK CODE END
@@ -1366,7 +1382,7 @@ namespace TrafficManager.Custom.PathFinding {
                     bool parkingAllowed = true;
 
                     // Parking AI: Determine if parking is allowed
-                    if (Options.parkingAI) {
+                    if (savedGameOptions_.parkingAI) {
                         if (isLogEnabled) {
                             DebugLog(
                                 unitId,
@@ -1528,7 +1544,7 @@ namespace TrafficManager.Custom.PathFinding {
                         }
                     } else {
                         // pocket car spawning
-                        if (Options.parkingAI) {
+                        if (savedGameOptions_.parkingAI) {
                             if (isLogEnabled) {
                                 DebugLog(
                                     unitId,
@@ -1792,11 +1808,11 @@ namespace TrafficManager.Custom.PathFinding {
                             unitId,
                             item,
                             "ProcessItemMain: vehicle -> vehicle: Custom routing\n" +
-                            $"\tOptions.advancedAI={Options.advancedAI}\n" +
+                            $"\tOptions.advancedAI={savedGameOptions_.advancedAI}\n" +
                             $"\tprevIsRouted={prevIsRouted}\n" +
                             $"\tm_isRoadVehicle={isRoadVehicle_}\n" +
                             $"\tprevIsCarLane={prevIsCarLane}\n" +
-                            $"\tm_stablePath={Options.advancedAI}");
+                            $"\tm_stablePath={savedGameOptions_.advancedAI}");
                     }
 
                     // NON-STOCK CODE START
@@ -1812,7 +1828,7 @@ namespace TrafficManager.Custom.PathFinding {
                      * Calculate Advanced Vehicle AI cost factors
                      * =============================================================================
                      */
-                    if (Options.advancedAI
+                    if (savedGameOptions_.advancedAI
                         && prevIsRouted
                         && isRoadVehicle_
                         && prevIsCarLane )
@@ -2630,12 +2646,6 @@ namespace TrafficManager.Custom.PathFinding {
                     $"\toffsetLength={offsetLength}");
             }
 #endif
-            if ((vehicleCategory_ & VehicleInfo.VehicleCategory.PublicTransportRoad) != 0 &&
-                (vehicleCategory_ & ~(VehicleInfo.VehicleCategory.Bus | VehicleInfo.VehicleCategory.Trolleybus | VehicleInfo.VehicleCategory.Taxi)) == 0)
-            {
-                offsetLength *= 0.75f;
-            }
-
             float baseLength = offsetLength / (prevLaneSpeed * maxLength_); // NON-STOCK CODE
             float comparisonValue = item.ComparisonValue; // NON-STOCK CODE
 #if ROUTING
@@ -2692,7 +2702,7 @@ namespace TrafficManager.Custom.PathFinding {
                 globalConf_.PathFinding.PublicTransportTransitionMaxPenalty;
 
             bool applyTransportTransferPenalty =
-                    Options.realisticPublicTransport &&
+                    savedGameOptions_.realisticPublicTransport &&
                     !stablePath_ &&
                     (allowedLaneTypes &
                      (NetInfo.LaneType.PublicTransport | NetInfo.LaneType.Pedestrian)) ==
@@ -2708,7 +2718,7 @@ namespace TrafficManager.Custom.PathFinding {
                     nextSegmentId,
                     $"ProcessItemCosts: Shall apply transport transfer penalty?\n" +
                     $"\tapplyTransportTransferPenalty={applyTransportTransferPenalty}\n" +
-                    $"\tOptions.realisticPublicTransport={Options.realisticPublicTransport}\n" +
+                    $"\tOptions.realisticPublicTransport={savedGameOptions_.realisticPublicTransport}\n" +
                     $"\tallowedLaneTypes={allowedLaneTypes}\n" +
                     $"\tallowedVehicleTypes={allowedVehicleTypes}\n" +
                     $"\tconf.pf.PubTranspTransitionMinPenalty={pfPublicTransportTransitionMinPenalty}\n" +
@@ -3130,6 +3140,11 @@ namespace TrafficManager.Custom.PathFinding {
                                 $"ProcessItemCosts: Adding next item\n\tnextItem={nextItem}");
                         }
 
+                        if ((nextLaneInfo.vehicleCategory & VehicleInfo.VehicleCategory.PublicTransportRoad) != 0 &&
+                            (nextLaneInfo.vehicleCategory & ~(VehicleInfo.VehicleCategory.Bus | VehicleInfo.VehicleCategory.Trolleybus | VehicleInfo.VehicleCategory.Taxi)) == 0) {
+                            nextItem.ComparisonValue -= baseLength * 0.25f;
+                        }
+
                         AddBufferItem(
 #if DEBUG
                             isLogEnabled,
@@ -3239,12 +3254,12 @@ namespace TrafficManager.Custom.PathFinding {
             // NON-STOCK CODE START
             bool mayCross = true;
 #if JUNCTIONRESTRICTIONS || CUSTOMTRAFFICLIGHTS
-            if (Options.junctionRestrictionsEnabled || Options.timedLightsEnabled) {
+            if (savedGameOptions_.junctionRestrictionsEnabled || savedGameOptions_.timedLightsEnabled) {
                 bool nextIsStartNode = nextNodeId == nextSegment.m_startNode;
 
                 if (nextIsStartNode || nextNodeId == nextSegment.m_endNode) {
 #if JUNCTIONRESTRICTIONS
-                    if (Options.junctionRestrictionsEnabled &&
+                    if (savedGameOptions_.junctionRestrictionsEnabled &&
                         item.Position.m_segment == nextSegmentId) {
                         // check if pedestrians are not allowed to cross here
                         if (!junctionManager.IsPedestrianCrossingAllowed(
@@ -3265,7 +3280,7 @@ namespace TrafficManager.Custom.PathFinding {
 #endif
 
 #if CUSTOMTRAFFICLIGHTS
-                    if (Options.timedLightsEnabled) {
+                    if (savedGameOptions_.timedLightsEnabled) {
                         // check if pedestrian light won't change to green
                         CustomSegmentLights lights = customTrafficLightsManager.GetSegmentLights(nextSegmentId, nextIsStartNode, false);
 
@@ -3606,7 +3621,7 @@ namespace TrafficManager.Custom.PathFinding {
              * =====================================================================================
              */
             bool canUseLane = CanUseLane(prevSegmentId, prevSegmentInfo, prevLaneIndex, prevLaneInfo, item.LaneId);
-            if (!canUseLane && Options.vehicleRestrictionsAggression ==
+            if (!canUseLane && savedGameOptions_.vehicleRestrictionsAggression ==
                 VehicleRestrictionsAggression.Strict) {
                 // vehicle is strictly prohibited to use this lane
                 if (isLogEnabled) {
@@ -3654,7 +3669,7 @@ namespace TrafficManager.Custom.PathFinding {
                 bool isStockUturnPoint = (nextNode.m_flags & (NetNode.Flags.End | NetNode.Flags.OneWayOut)) != NetNode.Flags.None;
 
 #if JUNCTIONRESTRICTIONS
-                if (Options.junctionRestrictionsEnabled) {
+                if (savedGameOptions_.junctionRestrictionsEnabled) {
                     bool nextIsStartNode = nextNodeId == prevSegment.m_startNode;
                     bool prevIsOutgoingOneWay =
                         nextIsStartNode
@@ -3712,7 +3727,7 @@ namespace TrafficManager.Custom.PathFinding {
             if (!canUseLane) {
                 laneSelectionCost *=
                     VehicleRestrictionsManager.PATHFIND_PENALTIES[
-                        (int)Options.vehicleRestrictionsAggression];
+                        (int)savedGameOptions_.vehicleRestrictionsAggression];
                 if (isLogEnabled) {
                     DebugLog(
                         unitId,
@@ -3727,7 +3742,7 @@ namespace TrafficManager.Custom.PathFinding {
              * Apply costs for large vehicles using inner lanes on highways
              * =======================================================================================================
              */
-            if (Options.preferOuterLane &&
+            if (savedGameOptions_.preferOuterLane &&
                 isHeavyVehicle_ &&
                 isRoadVehicle_ &&
                 prevIsCarLane &&
@@ -3745,7 +3760,7 @@ namespace TrafficManager.Custom.PathFinding {
                         item,
                         "ProcessItemRouted: Heavy trucks prefer outer lanes on highways: Applied lane costs\n"
                         + "\t" + $"laneSelectionCost={laneSelectionCost}\n"
-                        + "\t" + $"Options.preferOuterLane={Options.preferOuterLane}\n"
+                        + "\t" + $"savedGameOptions_.preferOuterLane={savedGameOptions_.preferOuterLane}\n"
                         + "\t" + $"m_isHeavyVehicle={isHeavyVehicle_}\n"
                         + "\t" + $"m_isRoadVehicle={isRoadVehicle_}\n"
                         + "\t" + $"prevIsCarLane={prevIsCarLane}\n"
@@ -4073,7 +4088,7 @@ namespace TrafficManager.Custom.PathFinding {
                                 int laneIndex,
                                 NetInfo.Lane laneInfo,
                                 uint laneId) {
-            if (!Options.vehicleRestrictionsEnabled ||
+            if (!savedGameOptions_.vehicleRestrictionsEnabled ||
                 queueItem_.vehicleType == ExtVehicleType.None ||
                 queueItem_.vehicleType == ExtVehicleType.Tram ||
                 queueItem_.vehicleType == ExtVehicleType.Trolleybus ||
@@ -4088,15 +4103,36 @@ namespace TrafficManager.Custom.PathFinding {
                 return true;
             }
 
-            ExtVehicleType allowedTypes = vehicleRestrictionsManager.GetAllowedVehicleTypes(
+            ExtVehicleType allowedLaneVehicleTypes = vehicleRestrictionsManager.GetAllowedVehicleTypes(
                 segmentId,
                 segmentInfo,
                 (uint)laneIndex,
                 laneInfo,
                 VehicleRestrictionsMode.Configured);
-
-            return (allowedTypes & queueItem_.vehicleType) != ExtVehicleType.None;
+            return (allowedLaneVehicleTypes & allowedPathVehicleTypes_) != ExtVehicleType.None;
         }
+
+        /// <summary>
+        /// Maps all allowed vehicle types to mixed ExtVehicleType
+        /// Later used to correctly test lane for match with configured allowed vehicle types
+        /// </summary>
+        /// <param name="vehicleTypes">Allowed additional vehicle types</param>
+        /// <returns>Combined ExtVehicleType of all alowed cargo vehicle lane types</returns>
+        private ExtVehicleType VehicleTypesToMixedCargoExtVehicle(VehicleInfo.VehicleType vehicleTypes) {
+            ExtVehicleType extVehicleType = ExtVehicleType.None;
+            if ((vehicleTypes & VehicleInfo.VehicleType.Train) != VehicleInfo.VehicleType.None) {
+                extVehicleType |= ExtVehicleType.CargoTrain;
+            }
+            if ((vehicleTypes & VehicleInfo.VehicleType.Plane) != VehicleInfo.VehicleType.None) {
+                extVehicleType |= ExtVehicleType.CargoPlane;
+            }
+            if ((vehicleTypes & VehicleInfo.VehicleType.Ship) != VehicleInfo.VehicleType.None) {
+                extVehicleType |= ExtVehicleType.CargoShip;
+            }
+
+            return extVehicleType;
+        }
+
 #endif
 
 #if ADVANCEDAI && ROUTING