+++ TLM/TLM/Manager/Impl/LaneArrowManager.cs
@@ -229,7 +229,7 @@ namespace TrafficManager.Manager.Impl {
 
         public override void OnLevelLoading() {
             base.OnLevelLoading();
-            if (Options.DedicatedTurningLanes) {
+            if (SavedGameOptions.Instance.DedicatedTurningLanes) {
                 // update dedicated turning lanes after patch has been applied.
                 UpdateDedicatedTurningLanePolicy(false);
             }