+++ TLM/TLM/Patch/_RoadBaseAI/UpdateNodePatch.cs
@@ -13,7 +13,7 @@ namespace TrafficManager.Patch._RoadBaseAI {
         [UsedImplicitly]
         public static void Postfix(ushort nodeID, ref NetNode data) {
 
-            if (Options.automaticallyAddTrafficLightsIfApplicable)
+            if (SavedGameOptions.Instance.automaticallyAddTrafficLightsIfApplicable)
             {
                 return;
             }