+++ TLM/TLM/State/OptionsTabs/PoliciesTab_AtJunctionsGroup.cs
@@ -9,46 +9,46 @@ namespace TrafficManager.State {
     public static class PoliciesTab_AtJunctionsGroup {
 
         public static CheckboxOption AllRelaxed =
-            new (nameof(Options.allRelaxed), Options.PersistTo.Savegame) {
+            new (nameof(SavedGameOptions.allRelaxed), Scope.Savegame) {
                 Label = "VR.Checkbox:All vehicles may ignore lane arrows",
                 Validator = DebugOnlyValidator,
             };
         public static CheckboxOption RelaxedBusses =
-            new (nameof(Options.relaxedBusses), Options.PersistTo.Savegame) {
+            new (nameof(SavedGameOptions.relaxedBusses), Scope.Savegame) {
                 Label = "VR.Checkbox:Buses may ignore lane arrows",
             };
         public static CheckboxOption AllowEnterBlockedJunctions =
-            new (nameof(Options.allowEnterBlockedJunctions), Options.PersistTo.Savegame) {
+            new (nameof(SavedGameOptions.allowEnterBlockedJunctions), Scope.Savegame) {
                 Label = "VR.Checkbox:Vehicles may enter blocked junctions",
                 Handler = OnJunctionRestrictionPolicyChanged,
             };
         public static CheckboxOption AllowUTurns =
-            new (nameof(Options.allowUTurns), Options.PersistTo.Savegame) {
+            new (nameof(SavedGameOptions.allowUTurns), Scope.Savegame) {
                 Label = "VR.Checkbox:Vehicles may do u-turns at junctions",
                 Handler = OnJunctionRestrictionPolicyChanged,
             };
         public static CheckboxOption AllowNearTurnOnRed =
-            new (nameof(Options.allowNearTurnOnRed), Options.PersistTo.Savegame) {
+            new (nameof(SavedGameOptions.allowNearTurnOnRed), Scope.Savegame) {
                 Label = "VR.Checkbox:Vehicles may turn on red",
                 Handler = OnJunctionRestrictionPolicyChanged,
             };
         public static CheckboxOption AllowFarTurnOnRed =
-            new (nameof(Options.allowFarTurnOnRed), Options.PersistTo.Savegame) {
+            new (nameof(SavedGameOptions.allowFarTurnOnRed), Scope.Savegame) {
                 Label = "VR.Checkbox:Also apply to left/right turns between one-way streets",
                 Indent = true,
                 Handler = OnJunctionRestrictionPolicyChanged,
             };
         public static CheckboxOption AllowLaneChangesWhileGoingStraight =
-            new (nameof(Options.allowLaneChangesWhileGoingStraight), Options.PersistTo.Savegame) {
+            new (nameof(SavedGameOptions.allowLaneChangesWhileGoingStraight), Scope.Savegame) {
                 Label = "VR.Checkbox:Vehicles going straight may change lanes at junctions",
                 Handler = OnJunctionRestrictionPolicyChanged,
             };
         public static CheckboxOption TrafficLightPriorityRules =
-            new (nameof(Options.trafficLightPriorityRules), Options.PersistTo.Savegame) {
+            new (nameof(SavedGameOptions.trafficLightPriorityRules), Scope.Savegame) {
                 Label = "VR.Checkbox:Vehicles follow priority rules at junctions with timedTL",
             };
         public static CheckboxOption AutomaticallyAddTrafficLightsIfApplicable =
-            new (nameof(Options.automaticallyAddTrafficLightsIfApplicable), Options.PersistTo.Savegame) {
+            new (nameof(SavedGameOptions.automaticallyAddTrafficLightsIfApplicable), Scope.Savegame) {
                 Label = "VR.Checkbox:Automatically add traffic lights if applicable",
             };
         public static CheckboxOption DedicatedTurningLanes =