+++ CHANGELOG.md
@@ -29,6 +29,28 @@ This changelog includes all versions and major variants of the mod going all the
 
 > Date format: dd/mm/yyyy
 
+#### TM:PE V[11.7.3.0](https://github.com/CitiesSkylinesMods/TMPE/compare/11.7.2.0...11.7.3.0) STABLE, 13/12/2022
+
+- [Meta] Compatibility patch for the game update 1.16.0-f3 (krzychu124)
+- [New] Improved bus/trolleybus stop path (pull to side earlier and use more of available space) #1688, #1690 (kianzarrin)
+- [Fixed] Cargo truck pathfinding via CargoVehicle networks (cargo stations/terminals) #1701, #1703 (krzychu124)
+- [Fixed] Lane connector on overlapping lanes #1706 (kianzarrin)
+- [Fixed] Parking signs on lanes with bus stop #1707 (kianzarrin)
+- [Updated] Minor changes in debug tools #1696, #1697 (kianzarrin)
+- [Updated] TMPE API Updates compatibility with other mods #1689 #1692 (kianzarrin)
+- [Steam] [TM:PE v11 STABLE](https://steamcommunity.com/sharedfiles/filedetails/?id=1637663252)
+
+#### TM:PE V11.7.3.0 TEST, 13/12/2022
+
+- [Meta] Compatibility patch for the game update 1.16.0-f3 (krzychu124)
+- [New] Improved bus/trolleybus stop path (pull to side earlier and use more of available space) #1688, #1690 (kianzarrin)
+- [Fixed] Cargo truck pathfinding via CargoVehicle networks (cargo stations/terminals) #1701, #1703 (krzychu124)
+- [Fixed] Lane connector on overlapping lanes #1706 (kianzarrin)
+- [Fixed] Parking signs on lanes with bus stop #1707 (kianzarrin)
+- [Updated] Minor changes in debug tools #1696, #1697 (kianzarrin)
+- [Updated] TMPE API Updates compatibility with other mods #1689 #1692 (kianzarrin)
+- [Steam] [TM:PE v11 TEST](https://steamcommunity.com/sharedfiles/filedetails/?id=2489276785)
+
 #### TM:PE V[11.7.2.0](https://github.com/CitiesSkylinesMods/TMPE/compare/11.7.1.2...11.7.2.0) STABLE, 15/11/2022
 
 - [Meta] Compatibility patch for the game update 1.15.1-f4 (krzychu124)